components/image_generator/requirements.txt
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=0.19.0
pymongo>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0 